+++ CHANGELOG.md
@@ -5,9 +5,10 @@
 ### v0.4.0
 
 - New option: `preload`, boolean:
-  - Enables higher priority to loading symbols;
-  - Disabled by default;
-  - More information: https://developer.mozilla.org/en-US/docs/Web/HTML/Attributes/rel/preload
+  - Enables higher priority for loading symbols;
+  - Disabled by default; More information:
+    - https://developer.mozilla.org/en-US/docs/Web/HTML/Attributes/rel/preload
+    - https://developer.mozilla.org/en-US/docs/Web/HTML/Attributes/rel/preconnect
 
 ### v0.3.0